Modify style of header/hero and of page body
Header/hero will now have a set width and elements of website will be
hidden when user adjusts the screen size. Furthermore, body of the
landing page was updated to a flex layout and footer was adjusted to
take up remaining space of landing page when it is zoomed out (instead
of leaving white space underneath the footer).